+++ webclient/src/components/header.ts
@@ -3,9 +3,9 @@ import {title} from './header.module.css';
 
 const Header = `<header style="padding:1em;width:100%">
     <div class="${title}">
-        <span style="font-weight: bold;">Diagram</span>
+        <span style="font-weight:bold">Diagram</span>
         <span> as </span>
-        <span style="font-weight: bold; font-style: italic;">Text</span>
+        <span style="font-weight:bold;font-style:italic">Text</span>
     </div>
 </header>`