+++ webclient/src/components/mailto.ts
@@ -1,0 +1,3 @@
+export default function MailToLinkStr(innerHTML: string = ""): string {
+    return `<a href="mailto:contact@diagramastext.dev" target="_blank" rel="noopener">${innerHTML}</a>`;
+}